backend/requirements.txt
# Core Framework (October 2025 LTS)
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
redis==5.0.8
pydantic==2.9.2
pydantic-settings==2.5.2

# Agent Frameworks
langgraph==0.2.20
langchain-core==0.3.10
langchain-community==0.3.2
crewai==0.55.1
crewai-tools==0.12.0

# LLM Integrations
langchain-openai==0.2.2
langchain-anthropic==0.3.4
langchain-groq==0.2.1

# Vector Store
qdrant-client==1.11.0

# Security & Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.2

# Monitoring
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp==1.27.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
httpx==0.27.2
faker==30.3.0

